Correct testing with Makefile
